Moved Firebase loading and saving functions into its own manager, no longer mixing DepositoryManager and Firebase things
